# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Collect&Verything",
        "version": "0.1"
    },
    "paths": {
        "/api/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get account information",
                "description": "Returns account data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Account not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Delete an account",
                "description": "Soft delete an account",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Patch an account",
                "description": "Update an account",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name"
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "male, female, other"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone number"
                    },
                    {
                        "name": "birthday",
                        "in": "query",
                        "description": "YYYY-MM-DD example : 1995-07-16",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email"
                    },
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "fr or en"
                    },
                    {
                        "name": "keep_logging",
                        "in": "query",
                        "description": "If the account stay logging"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/addresses/{id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get address information",
                "description": "Returns address data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Account not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete a address",
                "description": "Soft delete a address",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Patch a address",
                "description": "Update an address",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address id",
                        "required": true
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Address title"
                    },
                    {
                        "name": "address_line_1",
                        "in": "query",
                        "description": "Address line 1"
                    },
                    {
                        "name": "address_line_2",
                        "in": "query",
                        "description": "Address line 2"
                    },
                    {
                        "name": "zip_code",
                        "in": "query",
                        "description": "Zip code"
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City"
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Address country"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/admins/{id}": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Get admin information",
                "description": "Returns admin data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Admin not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admins"
                ],
                "summary": "Delete a admin",
                "description": "Soft delete a admin",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Admin id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admins"
                ],
                "summary": "Patch a admin",
                "description": "Update a admin",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Admin id"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "SUPER_ADMIN or ADMIN"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/carts/{id}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get cart information",
                "description": "Returns cart data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Cart id",
                        "required": true
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Cart not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "Delete a cart",
                "description": "Soft delete a cart",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Cart id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Carts"
                ],
                "summary": "Update a cart",
                "description": "Update a cart",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Cart id",
                        "required": true
                    },
                    {
                        "name": "payment_method",
                        "in": "query",
                        "description": "Payment method",
                        "required": false
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of cart",
                        "required": false
                    },
                    {
                        "name": "refund_at",
                        "in": "query",
                        "description": "Refund",
                        "required": false
                    },
                    {
                        "name": "paid_at",
                        "in": "query",
                        "description": "PAaid date",
                        "required": false
                    },
                    {
                        "name": "capture_at",
                        "in": "query",
                        "description": "Capture date",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart Updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/carts/users/{user_id}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get all carts information by Users",
                "description": "Returns cart data",
                "operationId": "listUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/carts/stores/{store_id}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get all carts information by store",
                "description": "Returns cart data",
                "operationId": "listStoreId",
                "parameters": [
                    {
                        "name": "store_id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee information",
                "description": "Returns employee data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Account not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete a employee",
                "description": "Soft delete a employee",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Employees"
                ],
                "summary": "Patch a employee",
                "description": "Update a employee",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Employee id"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "EMPLOYEE_ORDER or EMPLOYEE_STOCK"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/images/{id}": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get an image",
                "description": "Returns an image",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Image id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Account not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Delete an image",
                "description": "Soft delete an image",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Image id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/orders/users/{user_id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders information by Users",
                "description": "Returns order data",
                "operationId": "listUserId",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}